+++ hooks/useTtsAutoPlay.ts
@@ -2,13 +2,18 @@ import { useEffect, useRef } from 'react';
 
 export function useTtsAutoPlay(src: string | undefined, onReadyToRecord: () => void) {
   const audioRef = useRef<HTMLAudioElement | null>(null);
+  const callbackRef = useRef(onReadyToRecord);
+
+  useEffect(() => {
+    callbackRef.current = onReadyToRecord;
+  }, [onReadyToRecord]);
 
   useEffect(() => {
     if (!src) return;
     const el = audioRef.current;
     if (!el) return;
 
-    const ready = () => onReadyToRecord();
+    const ready = () => callbackRef.current();
 
     el.src = src;
     el.addEventListener('ended', ready);
@@ -21,7 +26,7 @@ export function useTtsAutoPlay(src: string | undefined, onReadyToRecord: () => v
       el.removeEventListener('ended', ready);
       el.removeEventListener('error', ready);
     };
-  }, [src, onReadyToRecord]);
+  }, [src]);
 
   return { audioRef } as const;
 }
